JI-872 IV support not being attached directly after it has been created
Problem: IV is initializing the video player and listening to different events
before it is attached to DOM. In some cases that means some code dependent
on IV being attached is run before it actually is.

I solved this by doing the relevant initialization when attach is invoked.

The indentation for this commit is intentionally wrong. In that way it is
possible to see what actually has changed for code reviewer. I will fix the
indentation in a separate commit.